+++ src/lib.rs
@@ -1,8 +1,8 @@
 use bvh::nalgebra::{Point2, Point3, Vector3};
 
-pub type Point2D = Point2<f64>;
-pub type Point = Point3<f64>;
-pub type Vector = Vector3<f64>;
+pub type Point2D = Point2<f32>;
+pub type Point = Point3<f32>;
+pub type Vector = Vector3<f32>;
 
 pub mod core;
 pub mod light;
